--- android/app/src/foss/java/chat/rocket/reactnative/CustomPushNotification.java
@@ -1,20 +1,0 @@
-package chat.rocket.reactnative;
-
-import android.content.Context;
-import android.os.Bundle;
-
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.wix.reactnativenotifications.core.AppLaunchHelper;
-import com.wix.reactnativenotifications.core.AppLifecycleFacade;
-import com.wix.reactnativenotifications.core.JsIOHelper;
-import com.wix.reactnativenotifications.core.notification.PushNotification;
-
-public class CustomPushNotification extends PushNotification {
-    public static ReactApplicationContext reactApplicationContext;
-
-    public CustomPushNotification(Context context, Bundle bundle, AppLifecycleFacade appLifecycleFacade, AppLaunchHelper appLaunchHelper, JsIOHelper jsIoHelper) {
-        super(context, bundle, appLifecycleFacade, appLaunchHelper, jsIoHelper);
-        reactApplicationContext = new ReactApplicationContext(context);
-    }
-
-}
