--- android/app/src/foss/java/chat/rocket/reactnative/DismissNotification.java
@@ -1,12 +1,0 @@
-package chat.rocket.reactnative;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-
-public class DismissNotification extends BroadcastReceiver {
-	@Override
-	public void onReceive(Context context, Intent intent) {
-
-	}
-}
