--- android/app/src/foss/java/chat/rocket/reactnative/ReplyBroadcast.java
@@ -1,13 +1,0 @@
-package chat.rocket.reactnative;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-
-public class ReplyBroadcast extends BroadcastReceiver {
-
-    @Override
-    public void onReceive(Context context, Intent intent) {
-
-    }
-}
