+++ android/app/src/main/AndroidManifest.xml
@@ -45,15 +45,6 @@
 			</intent-filter>
 		</activity>
 		<activity android:name="com.facebook.react.devsupport.DevSettingsActivity" />
-		<receiver
-			android:name=".ReplyBroadcast"
-			android:enabled="true"
-			android:exported="false" />
-		<receiver
-			android:name=".DismissNotification"
-			android:enabled="true"
-			android:exported="false" >
-		</receiver>
 		<activity
 				android:noHistory="true"
 				android:name=".share.ShareActivity"
