+++ android/app/src/main/java/chat/rocket/reactnative/MainApplication.java
@@ -1,8 +1,6 @@
 package chat.rocket.reactnative;
 
 import android.app.Application;
-import android.content.Context;
-import android.os.Bundle;
 
 import androidx.annotation.Nullable;
 
@@ -14,11 +12,6 @@ import com.facebook.soloader.SoLoader;
 import com.nozbe.watermelondb.WatermelonDBPackage;
 import com.reactnativecommunity.viewpager.RNCViewPagerPackage;
 import com.wix.reactnativekeyboardinput.KeyboardInputPackage;
-import com.wix.reactnativenotifications.core.AppLaunchHelper;
-import com.wix.reactnativenotifications.core.AppLifecycleFacade;
-import com.wix.reactnativenotifications.core.JsIOHelper;
-import com.wix.reactnativenotifications.core.notification.INotificationsApplication;
-import com.wix.reactnativenotifications.core.notification.IPushNotification;
 
 import org.unimodules.adapters.react.ModuleRegistryAdapter;
 import org.unimodules.adapters.react.ReactModuleRegistryProvider;
@@ -28,7 +21,7 @@ import java.util.List;
 
 import chat.rocket.reactnative.generated.BasePackageList;
 
-public class MainApplication extends Application implements ReactApplication, INotificationsApplication {
+public class MainApplication extends Application implements ReactApplication {
 
   private final ReactModuleRegistryProvider mModuleRegistryProvider = new ReactModuleRegistryProvider(new BasePackageList().getPackageList(), null);
 
@@ -74,16 +67,5 @@ public class MainApplication extends Application implements ReactApplication, IN
   public void onCreate() {
     super.onCreate();
     SoLoader.init(this, /* native exopackage */ false);
-  }
-
-  @Override
-  public IPushNotification getPushNotification(Context context, Bundle bundle, AppLifecycleFacade defaultFacade, AppLaunchHelper defaultAppLaunchHelper) {
-      return new CustomPushNotification(
-              context,
-              bundle,
-              defaultFacade,
-              defaultAppLaunchHelper,
-              new JsIOHelper()
-      );
   }
 }
