+++ android/app/src/play/AndroidManifest.xml
@@ -1,48 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
 		  xmlns:tools="http://schemas.android.com/tools"
 	package="chat.rocket.reactnative">
 
-	<uses-permission android:name="android.permission.INTERNET" />
-	<uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
-	<uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
-
 	<application
-		android:name=".MainApplication"
+		android:name=".MainPlayApplication"
 		android:label="@string/app_name"
 		android:icon="@mipmap/ic_launcher"
 		android:theme="@style/AppTheme"
 		android:networkSecurityConfig="@xml/network_security_config"
+		tools:replace="android:name"
 	>
-		<activity
-      android:name="com.zoontek.rnbootsplash.RNBootSplashActivity"
-      android:theme="@style/BootTheme">
-      <intent-filter>
-        <action android:name="android.intent.action.MAIN" />
-        <category android:name="android.intent.category.LAUNCHER" />
-      </intent-filter>
-    </activity>
-		<activity
-			android:name=".MainActivity"
-			android:label="@string/app_name"
-			android:configChanges="keyboard|keyboardHidden|orientation|screenSize|uiMode"
-			android:launchMode="singleTask"
-			android:windowSoftInputMode="adjustResize"
-			android:exported="true">
-			<intent-filter>
-				<action android:name="android.intent.action.DOWNLOAD_COMPLETE"/>
-			</intent-filter>
-			<intent-filter android:label="@string/app_name">
-				<action android:name="android.intent.action.VIEW" />
-				<category android:name="android.intent.category.DEFAULT" />
-				<category android:name="android.intent.category.BROWSABLE" />
-				<data android:scheme="https" android:host="go.rocket.chat" />
-				<data android:scheme="https" android:host="jitsi.rocket.chat" />
-				<data android:scheme="rocketchat" android:host="room" />
-				<data android:scheme="rocketchat" android:host="auth" />
-				<data android:scheme="rocketchat" android:host="jitsi.rocket.chat" />
-			</intent-filter>
-		</activity>
-		<activity android:name="com.facebook.react.devsupport.DevSettingsActivity" />
 		<receiver
 			android:name=".ReplyBroadcast"
 			android:enabled="true"
@@ -52,20 +20,6 @@
 			android:enabled="true"
 			android:exported="false" >
 		</receiver>
-		<activity
-				android:noHistory="true"
-				android:name=".share.ShareActivity"
-				android:configChanges="keyboard|keyboardHidden|orientation|screenSize"
-				android:label="@string/share_extension_name"
-				android:screenOrientation="portrait"
-				android:theme="@style/AppTheme" >
-				<intent-filter>
-					<action android:name="android.intent.action.SEND" />
-					<action android:name="android.intent.action.SEND_MULTIPLE" />
-					<category android:name="android.intent.category.DEFAULT" />
-					<data android:mimeType="*/*" />
-				</intent-filter>
-		</activity>
 		<meta-data
 				android:name="com.bugsnag.android.API_KEY"
 				android:value="${BugsnagAPIKey}" />
