+++ android/app/src/play/java/chat/rocket/reactnative/MainPlayApplication.java
@@ -1,0 +1,23 @@
+package chat.rocket.reactnative;
+
+import android.content.Context;
+import android.os.Bundle;
+
+import com.wix.reactnativenotifications.core.AppLaunchHelper;
+import com.wix.reactnativenotifications.core.AppLifecycleFacade;
+import com.wix.reactnativenotifications.core.JsIOHelper;
+import com.wix.reactnativenotifications.core.notification.INotificationsApplication;
+import com.wix.reactnativenotifications.core.notification.IPushNotification;
+
+public class MainPlayApplication extends MainApplication implements INotificationsApplication {
+    @Override
+    public IPushNotification getPushNotification(Context context, Bundle bundle, AppLifecycleFacade defaultFacade, AppLaunchHelper defaultAppLaunchHelper) {
+        return new CustomPushNotification(
+                context,
+                bundle,
+                defaultFacade,
+                defaultAppLaunchHelper,
+                new JsIOHelper()
+        );
+    }
+}
