+++ android/build.gradle
@@ -1,5 +1,12 @@
+def safeExtGet(prop, fallback) {
+    rootProject.ext.has(prop) ? rootProject.ext.get(prop) : fallback
+}
+
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
+    def taskRequests = getGradle().getStartParameter().getTaskRequests().toString().toLowerCase()
+    def isPlay = !taskRequests.contains("foss")
+
     ext {
         buildToolsVersion = "29.0.2"
         minSdkVersion = 23
@@ -8,7 +15,11 @@ buildscript {
         glideVersion = "4.9.0"
         kotlin_version = "1.3.50"
         supportLibVersion = "28.0.0"
+        libre_build = !(isPlay.toBoolean())
+        jitsi_url = isPlay ? "https://github.com/RocketChat/jitsi-maven-repository/raw/master/releases" : "https://github.com/RocketChat/jitsi-maven-repository/raw/libre/releases"
+        jitsi_version = isPlay ? "+" : "2.10.0-libre"
     }
+
     repositories {
         mavenLocal()
         google()
@@ -17,9 +28,6 @@ buildscript {
             url 'https://maven.fabric.io/public'
         }
     }
-
-    def taskRequests = getGradle().getStartParameter().getTaskRequests().toString().toLowerCase()
-    def isPlay = !taskRequests.contains("foss")
 
     dependencies {
         if (isPlay) {
@@ -40,7 +48,11 @@ allprojects {
             url("$rootDir/../node_modules/react-native/android")
         }
         maven {
-            url "https://github.com/RocketChat/jitsi-maven-repository/raw/master/releases"
+            // Android JSC is installed from npm
+            url("$rootDir/../node_modules/jsc-android/dist")
+        }
+        maven {
+            url safeExtGet("jitsi_url", "https://github.com/RocketChat/jitsi-maven-repository/raw/master/releases")
         }
 
         google()
